FR_716_smoke test
- improved existing smoke tests.
